refactor: move VmDetail's LastMessage to separate file
