+++ src/components/VmDetail/index.js
@@ -17,7 +17,6 @@ import {
 
 import { isWindows, templateNameRenderer, hrefWithoutHistory } from '../../helpers'
 
-import Time from '../Time'
 import FieldHelp from '../FieldHelp/index'
 import DetailContainer from '../DetailContainer'
 import ConsoleOptions from '../ConsoleOptions/index'
@@ -25,35 +24,11 @@ import VmDisks from '../VmDisks/index'
 import VmsListNavigation from '../VmsListNavigation/index'
 import VmStatus from './VmStatus'
 import { NextRunLabel, OptimizedForLabel } from './labels'
+import LastMessage from './LastMessage'
 
 import { canConsole, userFormatOfBytes, VmIcon } from 'ovirt-ui-components'
 import Selectors from '../../selectors'
 import ConsoleButton from '../VmActions/ConsoleButton'
-
-const LastMessage = ({ vmId, userMessages }) => {
-  const vmMessages = userMessages.get('records')
-    .filter(msg => (msg.failedAction && msg.failedAction.payload && msg.failedAction.payload.vmId === vmId))
-    .sort((msg1, msg2) => (msg1.time - msg2.time))
-
-  const lastMessage = vmMessages.last()
-
-  if (!lastMessage) {
-    return null
-  }
-
-  return (
-    <span>
-      <Time time={lastMessage.time} />
-      <pre>
-        {lastMessage.message}
-      </pre>
-    </span>
-  )
-}
-LastMessage.propTypes = {
-  vmId: PropTypes.string.isRequired,
-  userMessages: PropTypes.object.isRequired,
-}
 
 const VmConsoles = ({ vm, onConsole, onRDP }) => {
   const vmConsoles = vm.get('consoles').valueSeq()
